// src/core/bindings.d.ts
export declare const BINDINGS: {
    FACTORY: symbol;
    CLI: symbol;
    COMMANDS: symbol;
    COMMANDS_FACTORY: symbol;
    LOG: symbol;
    CONFIG: symbol;
    INPUT: symbol;
    OUTPUT: symbol;
    DESCRIPTOR: symbol;
    ROOT_DEFINITION: symbol;
    GLOBAL_DEFINITION: symbol;
    ROOT_DEFINITION_PARSER_FACTORY: symbol;
    OPTIONS_DEFINITION: symbol;
    COMMANDS_DEFINITION: symbol;
    ARGUMENTS_DEFINITION: symbol;
    OPTIONS_DEFINITION_PARSER: symbol;
    COMMANDS_DEFINITION_PARSER: symbol;
    ARGUMENTS_DEFINITION_PARSER: symbol;
    PARSED_OPTIONS_DEFINITION: symbol;
    PARSED_ARGUMENTS_DEFINITION: symbol;
    PARSED_COMMANDS_DEFINITION: symbol;
    DEFINITION_SIGNATURE_PARSER: symbol;
    OPTIONS_DEFINITION_PARSER_FACTORY: symbol;
    ARGUMENTS_DEFINITION_PARSER_FACTORY: symbol;
    COMMANDS_DEFINITION_PARSER_FACTORY: symbol;
    HELP_WRITER: symbol;
};
export default BINDINGS;